backend/requirements.txt
anthropic==0.94.1
requests==2.32.5
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
pyjwt==2.12.1
bcrypt==4.2.0
python-dotenv==1.0.1
cryptography==44.0.1
slowapi==0.1.9
bleach==6.1.0
python-multipart==0.0.12
pypdf==4.1.0
python-docx==1.1.2
Pillow==11.3.0
